+++ src/assets/images/RoboticArm/arm.jpeg
