+++ src/assets/images/RoboticArm/list.png
