+++ src/assets/images/RoboticArm/wiring.png
