+++ src/assets/images/T6Drone/complete.jpg
