+++ src/assets/images/T6Drone/model.jpg
