+++ src/plugins/vuetify.ts
@@ -4,7 +4,7 @@ import { aliases, mdi } from 'vuetify/iconsets/mdi'
 import { fa } from 'vuetify/iconsets/fa'
 import '@mdi/font/css/materialdesignicons.css'
 import '@fortawesome/fontawesome-free/css/all.css'
-import '../css/style.scss'
+import '@/assets/scss/style.scss'
 import { VTreeview } from 'vuetify/lib/labs/components.mjs'
 
 const myTheme = {
