+++ vue.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+	transpileDependencies: true,
+
+	pluginOptions: {
+		vuetify: {},
+	},
+}
