--- vue.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from '@vue/cli-service'
-export default defineConfig({
-	transpileDependencies: true,
-
-	pluginOptions: {
-		vuetify: {
-			// https://github.com/vuetifyjs/vuetify-loader/tree/next/packages/vuetify-loader
-		},
-	},
-})
